chore: prune let.rs
